Intelligent Trolley System
One of the important aspects of modern electronic technology is
embedded systems based on micro controllers. The main aim of science
has always been to make our lives easier. In this paper, we discuss an
innovative concept of Radio Frequency Identification(RFID) which is
used to implement an Intelligent Shopping and billing trolley. Some of
the issues faced by customers while shopping are:
Wastage of valuable time while standing in long queues at the billing
counter.
No track of expenditure.

We observed that the main cause for long queues at the billing counter
is not the crowd but the time spent in scanning each and every item
using the Bar code Technology.
Our project consist of an RFID based trolley that communicates with the
billing side using a ZIGBEE(NRF24L01) module. The RFID reader scans all
the items as and when they are put in the trolley. The record of the
items bought is stored in the micro controller memory along with their
individual costs as well as the total expenditure.This information will
be displayed on a LCD screen which will also be placed on the trolley
for the customer to verify the item bought and to keep a track on the
amount spent on each item. At the billing side, the employee can get an
itemized bill from each and every trolley just by giving the trolley
number as the input to a software which would then print the itemized
bill. Data can be erased from the micro controller memory after the
bill is printed so as to make that trolley reusable.